Update public routes and coverage
