+++ src/data/resources.ts
@@ -7,7 +7,7 @@ export const resources: Resource[] = [
     description: "Comprehensive guidelines for community health workers on providing maternal healthcare support.",
     category: "education",
     type: "PDF",
-    image: "/images/resources/maternal-health.jpg",
+    image: "/highlights/Akomapa-62.jpg",
     url: "/documents/sample-resources/maternal-health-guidelines.pdf",
     downloadUrl: "/documents/sample-resources/maternal-health-guidelines.pdf",
     programs: ["community-clinics", "health-education"],
@@ -19,7 +19,7 @@ export const resources: Resource[] = [
     description: "Educational materials for teaching preventive healthcare practices in community settings.",
     category: "education",
     type: "Toolkit",
-    image: "/images/resources/preventive-care.jpg",
+    image: "/highlights/Akomapa-28.jpg",
     url: "/resources/preventive-healthcare-toolkit",
     downloadUrl: "/documents/sample-resources/preventive-healthcare-toolkit.zip",
     programs: ["health-education"],
@@ -31,7 +31,7 @@ export const resources: Resource[] = [
     description: "Research study examining the effectiveness of community clinics in improving rural health metrics.",
     category: "research",
     type: "Article",
-    image: "/images/resources/research-paper.jpg",
+    image: "/highlights/Akomapa-61.jpg",
     url: "https://example.com/research/community-clinics-impact",
     programs: ["community-clinics", "research"],
     date: "2023-01-10"
@@ -42,7 +42,7 @@ export const resources: Resource[] = [
     description: "Video training on basic life support techniques for healthcare workers in resource-limited settings.",
     category: "education",
     type: "Video",
-    image: "/images/resources/life-support.jpg",
+    image: "/highlights/Akomapa-20.jpg",
     url: "https://www.youtube.com/watch?v=example",
     programs: ["medical-training"],
     date: "2022-11-05"
@@ -53,7 +53,7 @@ export const resources: Resource[] = [
     description: "Guide for healthcare providers on managing chronic diseases in community settings.",
     category: "guide",
     type: "PDF",
-    image: "/images/resources/chronic-disease.jpg",
+    image: "/highlights/Akomapa-40.jpg",
     url: "/documents/sample-resources/chronic-disease-handbook.pdf",
     downloadUrl: "/documents/sample-resources/chronic-disease-handbook.pdf",
     programs: ["community-clinics", "medical-training"],
@@ -65,7 +65,7 @@ export const resources: Resource[] = [
     description: "Case studies of innovative approaches to healthcare delivery in rural and underserved areas.",
     category: "research",
     type: "Article",
-    image: "/images/resources/innovation.jpg",
+    image: "/highlights/Akomapa-10.jpg",
     url: "/resources/innovations-rural-healthcare",
     programs: ["research", "community-clinics"],
     date: "2022-12-15"
@@ -76,7 +76,7 @@ export const resources: Resource[] = [
     description: "Structured curriculum for training community members on essential health topics.",
     category: "education",
     type: "Toolkit",
-    image: "/images/resources/curriculum.jpg",
+    image: "/highlights/Akomapa-66.jpg",
     url: "/documents/sample-resources/health-education-curriculum.pdf",
     downloadUrl: "/documents/sample-resources/health-education-curriculum.pdf",
     programs: ["health-education"],
@@ -88,7 +88,7 @@ export const resources: Resource[] = [
     description: "Best practices for managing healthcare facilities in resource-constrained environments.",
     category: "guide",
     type: "PDF",
-    image: "/images/resources/facility-management.jpg",
+    image: "/highlights/ucc.jpg",
     url: "/documents/sample-resources/facility-management.pdf",
     downloadUrl: "/documents/sample-resources/facility-management.pdf",
     programs: ["community-clinics", "medical-training"],
@@ -100,7 +100,7 @@ export const resources: Resource[] = [
     description: "Visual educational materials on key maternal and child health topics for community education.",
     category: "education",
     type: "Infographic",
-    image: "/images/resources/infographics.jpg",
+    image: "/gallery/gallery-pic-5.jpg",
     url: "/resources/maternal-child-infographics",
     downloadUrl: "/documents/sample-resources/maternal-child-infographics.zip",
     programs: ["health-education", "community-clinics"],
